Consolidate shield banner code
