+++ shieldlib/src/types.ts
@@ -218,3 +218,8 @@ export interface GraphicsFactory {
    */
   pixelRatio(): number;
 }
+
+export interface Dimension {
+  width: number;
+  height: number;
+}
